requirements.txt
Django==4.1.1
djangorestframework==3.14.0
Pillow==9.2.0
psycopg2==2.9.1
channels==3.0.4
channels-redis==3.2.0
requests==2.26.0